login-cloud-function/requirements.txt
SQLAlchemy==1.3.12
pg8000==1.13.2
bcrypt==3.2.0
PyJWT==2.1.0